Task_02/requirements.txt
streamlit>=1.33.0
phidata>=2.3.4 
google-generativeai>=0.3.0
python-dotenv>=1.0.0